+++ type-generator/src/lib.rs
@@ -1,4 +1,7 @@
 mod dag;
+pub mod ir;
+mod to_tokens;
+mod transformer;
 
 use once_cell::sync::Lazy;
 use std::{
@@ -15,487 +18,12 @@ use swc_common::{
 
 use swc_ecma_parser::{lexer::Lexer, Capturing, Parser, StringInput, Syntax};
 
-use proc_macro2::{TokenStream, TokenTree};
-use quote::{quote, ToTokens, TokenStreamExt};
+use ir::*;
 
 macro_rules! id {
     ($($tt:tt)*) => {
         proc_macro2::Ident::new($($tt)*, proc_macro2::Span::call_site())
     };
-}
-
-pub enum RustType {
-    String,
-    Number,
-    Boolean,
-    Custom(TokenStream),
-    Array(Box<RustType>),
-    Empty, // ()
-    Unknown,
-    UnknownLiteral,
-    UnknownIntersection,
-    UnknownUnion,
-}
-
-impl RustType {
-    pub fn is_unknown(&self) -> bool {
-        match &self {
-            RustType::Unknown
-            | RustType::UnknownLiteral
-            | RustType::UnknownIntersection
-            | RustType::UnknownUnion => true,
-            RustType::Array(t) => t.is_unknown(),
-            _ => false,
-        }
-    }
-}
-
-pub struct RustStruct {
-    pub attr: RustContainerAttrs,
-    pub name: String,
-    pub member: Vec<RustStructMember>,
-}
-
-#[derive(Default)]
-pub enum RustContainerAttrs {
-    #[default]
-    Default,
-    With(Vec<RustStructAttr>),
-}
-
-impl RustContainerAttrs {
-    fn add_attr(&mut self, a: RustStructAttr) {
-        match self {
-            RustContainerAttrs::Default => *self = Self::With(vec![a]),
-            RustContainerAttrs::With(v) => v.push(a),
-        }
-    }
-    fn is_tagged_enum(&self) -> bool {
-        match self {
-            RustContainerAttrs::Default => false,
-            RustContainerAttrs::With(attrs) => attrs
-                .iter()
-                .filter_map(|attr| attr.as_serde())
-                .any(SerdeContainerAttr::is_tag),
-        }
-    }
-}
-
-impl ToTokens for RustFieldAttr {
-    fn to_tokens(&self, tokens: &mut TokenStream) {
-        tokens.extend(
-            match self {
-                RustFieldAttr::Serde(s) => quote! {
-                    #[serde(#s)]
-                },
-            }
-            .into_iter(),
-        )
-    }
-}
-
-pub enum RustStructAttr {
-    Serde(SerdeContainerAttr),
-}
-
-impl RustStructAttr {
-    pub fn as_serde(&self) -> Option<&SerdeContainerAttr> {
-        let Self::Serde(v) = self;
-        Some(v)
-    }
-}
-
-impl ToTokens for RustStructAttr {
-    fn to_tokens(&self, tokens: &mut TokenStream) {
-        tokens.extend(
-            match self {
-                RustStructAttr::Serde(s) => quote! {
-                    #[serde(#s)]
-                },
-            }
-            .into_iter(),
-        )
-    }
-}
-
-pub enum SerdeContainerAttr {
-    RenameAll(RenameRule),
-    Tag(String),
-}
-
-impl SerdeContainerAttr {
-    /// Returns `true` if the serde container attr is [`Tag`].
-    ///
-    /// [`Tag`]: SerdeContainerAttr::Tag
-    #[must_use]
-    pub fn is_tag(&self) -> bool {
-        matches!(self, Self::Tag(..))
-    }
-}
-
-impl ToTokens for SerdeContainerAttr {
-    fn to_tokens(&self, tokens: &mut TokenStream) {
-        tokens.extend(
-            match self {
-                SerdeContainerAttr::RenameAll(r) => {
-                    let r = r.to_string();
-                    quote! {
-                        rename_all = #r
-                    }
-                }
-                SerdeContainerAttr::Tag(name) => quote! {
-                    tag = #name
-                },
-            }
-            .into_iter(),
-        )
-    }
-}
-
-pub enum SerdeFieldAttr {
-    Rename(String),
-    Borrow,
-}
-
-impl ToTokens for SerdeFieldAttr {
-    fn to_tokens(&self, tokens: &mut TokenStream) {
-        tokens.extend(
-            match self {
-                SerdeFieldAttr::Rename(s) => quote! {
-                    rename = #s
-                },
-                SerdeFieldAttr::Borrow => quote! {
-                    borrow = "'a"
-                },
-            }
-            .into_iter(),
-        )
-    }
-}
-
-pub enum SerdeVariantAttr {
-    Rename(String),
-    Borrow {
-        /// lifetime parameter without `'`
-        lifetime: String,
-    },
-}
-
-pub enum RenameRule {
-    PascalCase,
-    SnakeCase,
-    ScreamingSnakeCase,
-}
-
-impl RenameRule {
-    /// Returns `true` if the rename rule is [`PascalCase`].
-    ///
-    /// [`PascalCase`]: RenameRule::PascalCase
-    #[must_use]
-    pub fn is_pascal_case(&self) -> bool {
-        matches!(self, Self::PascalCase)
-    }
-    fn convert_to_pascal(&self, s: &mut String) {
-        match self {
-            RenameRule::PascalCase => (),
-            RenameRule::SnakeCase | RenameRule::ScreamingSnakeCase => {
-                *s = s
-                    .split('_')
-                    .map(|term| {
-                        let mut term = term.to_ascii_lowercase();
-                        if let Some(c) = term.chars().next() {
-                            let capital_ch = c.to_ascii_uppercase();
-                            term.replace_range(..1, &capital_ch.to_string());
-                        }
-                        term
-                    })
-                    .collect::<Vec<_>>()
-                    .concat();
-            }
-        }
-    }
-}
-
-impl ToString for RenameRule {
-    fn to_string(&self) -> String {
-        match self {
-            RenameRule::PascalCase => "PascalCase",
-            RenameRule::SnakeCase => "snake_case",
-            RenameRule::ScreamingSnakeCase => "SCREAMING_SNAKE_CASE",
-        }
-        .to_string()
-    }
-}
-
-pub struct RustEnum {
-    pub attr: RustContainerAttrs,
-    pub name: String,
-    pub member: Vec<RustEnumMember>,
-}
-
-pub struct RustStructMember {
-    pub attr: RustFieldAttrs,
-    pub name: String,
-    pub ty: RustMemberType,
-    pub comment: Option<String>,
-}
-
-#[derive(Default)]
-pub enum RustFieldAttrs {
-    #[default]
-    Default,
-    With(Vec<RustFieldAttr>),
-}
-
-impl RustFieldAttrs {
-    fn add_attr(&mut self, a: RustFieldAttr) {
-        match self {
-            Self::Default => *self = Self::With(vec![a]),
-            Self::With(v) => v.push(a),
-        }
-    }
-}
-
-pub enum RustFieldAttr {
-    Serde(SerdeFieldAttr),
-}
-
-pub struct RustMemberType {
-    pub ty: RustType,
-    pub is_optional: bool,
-}
-
-impl RustMemberType {
-    pub fn is_unknown(&self) -> bool {
-        self.ty.is_unknown()
-    }
-}
-
-pub enum RustEnumMember {
-    Nullary(String),
-    /// has the same ident. this is unary
-    Unary(String),
-    UnaryNamed {
-        variant_name: String,
-        type_name: String,
-    },
-}
-
-impl RustEnumMember {
-    fn name_unary(&mut self, variant_name: String) {
-        match self {
-            RustEnumMember::Unary(u) => {
-                *self = Self::UnaryNamed {
-                    variant_name,
-                    type_name: u.clone(),
-                }
-            }
-            _ => unreachable!("do not call with this"),
-        }
-    }
-
-    /// Returns `true` if the rust enum member is [`Nullary`].
-    ///
-    /// [`Nullary`]: RustEnumMember::Nullary
-    #[must_use]
-    pub fn is_nullary(&self) -> bool {
-        matches!(self, Self::Nullary(..))
-    }
-
-    pub fn as_unary(&self) -> Option<&String> {
-        if let Self::Unary(v) = self {
-            Some(v)
-        } else {
-            None
-        }
-    }
-}
-
-impl ToTokens for RustType {
-    fn to_tokens(&self, tokens: &mut TokenStream) {
-        let s = match self {
-            RustType::String => "String",
-            RustType::Number => "usize",
-            RustType::Boolean => "bool",
-            RustType::Custom(s) => {
-                tokens.extend(s.clone().into_iter());
-                return;
-            }
-            RustType::Array(t) => {
-                tokens.extend(
-                    quote! {
-                        Vec<#t>
-                    }
-                    .into_iter(),
-                );
-                return;
-            }
-            RustType::Empty => {
-                tokens.append(TokenTree::Group(proc_macro2::Group::new(
-                    proc_macro2::Delimiter::Parenthesis,
-                    Default::default(),
-                )));
-                return;
-            }
-            RustType::Unknown => "Unknown",
-            RustType::UnknownLiteral => "UnknownLiteral",
-            RustType::UnknownIntersection => "UnknownIntersection",
-            RustType::UnknownUnion => "UnknownUnion",
-        };
-        tokens.append(TokenTree::Ident(id!(s)));
-    }
-}
-
-impl ToTokens for RustMemberType {
-    fn to_tokens(&self, tokens: &mut TokenStream) {
-        let inner_ty = &self.ty;
-        tokens.extend(
-            if self.is_optional {
-                quote! {
-                    Option<#inner_ty>
-                }
-            } else {
-                quote! {
-                    #inner_ty
-                }
-            }
-            .into_iter(),
-        )
-    }
-}
-
-impl ToTokens for RustStructMember {
-    fn to_tokens(&self, tokens: &mut TokenStream) {
-        let Self {
-            attr,
-            name,
-            ty,
-            comment,
-        } = self;
-        let name = id!(name);
-
-        tokens.extend(
-            if self.ty.is_unknown() {
-                quote! {
-                    /* unknown type */
-                }
-            } else {
-                let mut attrs = TokenStream::new();
-                if let Some(comment) = comment {
-                    attrs = quote! {
-                        #[doc=#comment]
-                    }
-                }
-                match attr {
-                    RustFieldAttrs::Default => (),
-                    RustFieldAttrs::With(w) => {
-                        attrs = quote! {
-                            #attrs
-                            #(#w)*
-                        }
-                    }
-                }
-                quote! {
-                    #attrs
-                    pub #name: #ty,
-                }
-            }
-            .into_iter(),
-        );
-    }
-}
-
-impl ToTokens for RustStruct {
-    fn to_tokens(&self, tokens: &mut TokenStream) {
-        let Self { name, member, attr } = self;
-        let name = id!(name);
-        tokens.extend(
-            match attr {
-                RustContainerAttrs::Default => quote! {
-                    #[derive(Debug, Deserialize)]
-                },
-                RustContainerAttrs::With(w) => quote! {
-                    #[derive(Debug, Deserialize)]
-                    #(#w)*
-                },
-            }
-            .into_iter(),
-        );
-
-        tokens.extend(
-            quote! {
-                pub struct #name {
-                    #(#member)*
-                }
-            }
-            .into_iter(),
-        );
-    }
-}
-
-impl ToTokens for RustEnum {
-    fn to_tokens(&self, tokens: &mut TokenStream) {
-        let Self { name, member, attr } = self;
-        let name = id!(name);
-        tokens.extend(
-            if attr.is_tagged_enum() || member.iter().all(|m| m.is_nullary()) {
-                quote! {
-                    #[derive(Debug, Deserialize)]
-                }
-            } else {
-                quote! {
-                    #[derive(Debug)]
-                }
-            }
-            .into_iter(),
-        );
-        match attr {
-            RustContainerAttrs::Default => (),
-            RustContainerAttrs::With(w) => {
-                tokens.extend(
-                    quote! {
-                        #(#w)*
-                    }
-                    .into_iter(),
-                );
-            }
-        }
-
-        tokens.extend(
-            quote! {
-                pub enum #name {
-                    #(#member)*
-                }
-            }
-            .into_iter(),
-        );
-    }
-}
-
-impl ToTokens for RustEnumMember {
-    fn to_tokens(&self, tokens: &mut TokenStream) {
-        tokens.extend(
-            match self {
-                RustEnumMember::Nullary(v) => {
-                    let v = id!(v);
-                    quote!(#v,)
-                }
-                RustEnumMember::Unary(a) => {
-                    let a = id!(a);
-                    quote!(#a(#a),)
-                }
-                RustEnumMember::UnaryNamed {
-                    variant_name,
-                    type_name,
-                } => {
-                    let variant_name = id!(variant_name);
-                    let type_name = id!(type_name);
-                    quote!(#variant_name(#type_name),)
-                }
-            }
-            .into_iter(),
-        )
-    }
 }
 
 fn interface2struct(
@@ -636,27 +164,7 @@ fn tunion2enum(name: &str, tunion: &swc_ecma_ast::TsUnionType) -> RustEnum {
     }
 }
 
-enum RustSegment {
-    Struct(RustStruct),
-    Enum(RustEnum),
-    Alias(proc_macro2::Ident, RustType),
-}
-
-impl RustSegment {
-    fn into_token_stream(self) -> TokenStream {
-        match self {
-            RustSegment::Struct(s) => s.into_token_stream(),
-            RustSegment::Enum(e) => e.into_token_stream(),
-            RustSegment::Alias(ident, typ) => quote! {
-                pub type #ident = #typ;
-            },
-        }
-    }
-}
-
-type LiteralKeyMap = HashMap<String, HashMap<String, String>>;
-
-pub fn dts2rs(dts_file: &PathBuf) -> TokenStream {
+pub fn dts2rs(dts_file: &PathBuf) -> proc_macro2::TokenStream {
     let module = extract_module(dts_file);
 
     let mut segments = Vec::new();
@@ -699,9 +207,9 @@ pub fn dts2rs(dts_file: &PathBuf) -> TokenStream {
                 let typ = &talias.type_ann;
                 match typ.as_ref() {
                     swc_ecma_ast::TsType::TsTypeRef(tref) => {
-                        let rhs = &tref.type_name.as_ident().unwrap().sym;
-                        let rhs = id!(rhs);
-                        RustSegment::Alias(ident, RustType::Custom(quote!(#rhs)))
+                        let rhs = tref.type_name.as_ident().unwrap().sym.as_ref();
+                        let rhs = rhs.to_owned();
+                        RustSegment::Alias(ident, RustType::Custom(rhs))
                     }
                     swc_ecma_ast::TsType::TsUnionOrIntersectionType(tuoi) => {
                         let tunion = tuoi.as_ts_union_type().unwrap();
@@ -733,137 +241,14 @@ pub fn dts2rs(dts_file: &PathBuf) -> TokenStream {
     }
 
     for segment in &mut segments {
-        adapt_internal_tag(segment, &lkm);
-        adapt_rename_all(segment);
+        transformer::adapt_internal_tag(segment, &lkm);
+        transformer::adapt_rename_all(segment);
     }
 
     segments
         .into_iter()
         .flat_map(|rss| rss.into_token_stream())
         .collect()
-}
-
-fn adapt_internal_tag(segment: &mut RustSegment, lkm: &LiteralKeyMap) -> Option<()> {
-    if let RustSegment::Enum(re) = segment {
-        let mut props: HashMap<String, String> = Default::default();
-        for memb in &re.member {
-            let inter = memb.as_unary()?;
-            let map = lkm.get(inter)?;
-            if props.is_empty() {
-                props = map.clone();
-                continue;
-            }
-            props.retain(|k, _| map.contains_key(k));
-            if props.is_empty() {
-                return None;
-            }
-        }
-        // assert !props.is_empty()
-        assert_eq!(props.len(), 1);
-        let tag_name = props.keys().next().unwrap().to_owned();
-        for memb in &mut re.member {
-            let inter = memb.as_unary().unwrap();
-            let variant_name = lkm.get(inter).unwrap().get(&tag_name).unwrap().to_owned();
-            memb.name_unary(variant_name);
-        }
-        re.attr
-            .add_attr(RustStructAttr::Serde(SerdeContainerAttr::Tag(tag_name)));
-    }
-    Some(())
-}
-
-fn adapt_rename_all(segment: &mut RustSegment) -> Option<()> {
-    #[derive(Debug, Clone, Copy, PartialEq)]
-    enum CaseConvention {
-        Lower,
-        Upper,
-        Snake,
-        Pascal,
-        ScreamingSnake,
-    }
-    use CaseConvention::*;
-    /// expect input follows one of above case
-    fn detect_case(s: &str) -> CaseConvention {
-        if s.starts_with(char::is_uppercase) {
-            if s.contains('_') {
-                ScreamingSnake
-            } else if s.chars().all(char::is_uppercase) {
-                Upper
-            } else {
-                Pascal
-            }
-        } else if s.contains('_') {
-            Snake
-        } else {
-            Lower
-        }
-    }
-    impl CaseConvention {
-        fn cast(&mut self, other: Self) -> Option<()> {
-            if self == &other {
-                return Some(());
-            }
-            match (&self, &other) {
-                (Lower, Snake) | (Upper, ScreamingSnake) => {
-                    *self = other;
-                }
-                (Snake, Lower) | (ScreamingSnake, Upper) => {
-                    // do nothing
-                }
-                _ => None?,
-            }
-            Some(())
-        }
-        fn into_rename_rule(self) -> RenameRule {
-            match self {
-                Lower | Snake => RenameRule::SnakeCase,
-                Pascal => RenameRule::PascalCase,
-                Upper | ScreamingSnake => RenameRule::ScreamingSnakeCase,
-            }
-        }
-    }
-    if let RustSegment::Enum(re) = segment {
-        let mut conv: Option<CaseConvention> = None;
-        for memb in &re.member {
-            let s = match memb {
-                RustEnumMember::Nullary(v) => v,
-                RustEnumMember::Unary(v) => v,
-                RustEnumMember::UnaryNamed { variant_name, .. } => variant_name,
-            };
-            match conv.as_mut() {
-                Some(conv) => {
-                    let new = detect_case(s);
-                    conv.cast(new)?
-                }
-                None => {
-                    conv = Some(detect_case(s));
-                }
-            };
-        }
-        let rr = conv?.into_rename_rule();
-        if rr.is_pascal_case() {
-            return None;
-        }
-        for memb in &mut re.member {
-            match memb {
-                RustEnumMember::Unary(v) => {
-                    let type_name = v.to_owned();
-                    rr.convert_to_pascal(v);
-                    *memb = RustEnumMember::UnaryNamed {
-                        variant_name: v.to_owned(),
-                        type_name,
-                    };
-                }
-                RustEnumMember::Nullary(variant_name)
-                | RustEnumMember::UnaryNamed { variant_name, .. } => {
-                    rr.convert_to_pascal(variant_name);
-                }
-            };
-        }
-        re.attr
-            .add_attr(RustStructAttr::Serde(SerdeContainerAttr::RenameAll(rr)));
-    }
-    Some(())
 }
 
 fn extract_module(dts_file: &PathBuf) -> swc_ecma_ast::Module {
@@ -983,8 +368,8 @@ fn ts_type_to_rs(typ: &swc_ecma_ast::TsType) -> (bool, RustType) {
         }
         swc_ecma_ast::TsType::TsLitType(_tslit) => RustType::UnknownLiteral,
         swc_ecma_ast::TsType::TsTypeRef(tref) => {
-            let id = id!(tref.type_name.as_ident().unwrap().sym.as_ref());
-            RustType::Custom(quote!(#id))
+            let id = tref.type_name.as_ident().unwrap().sym.as_ref();
+            RustType::Custom(id.to_owned())
         }
         swc_ecma_ast::TsType::TsArrayType(tarray) => {
             let (_n, etype) = ts_type_to_rs(&tarray.elem_type);
