+++ type-generator/src/transformer.rs
@@ -1,0 +1,5 @@
+mod internal_tag;
+mod rename_all;
+
+pub use internal_tag::adapt_internal_tag;
+pub use rename_all::adapt_rename_all;
